% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvalue.R
\name{base_sn}
\alias{base_sn}
\title{Stage 1 Critical Value Backend}
\usage{
base_sn(n, k, alpha)
}
\arguments{
\item{n}{the sample size.}

\item{k}{the number of moments.}

\item{alpha}{the confidence level required.}
}
\value{
a float for the critical value.
}
\description{
Stage 1 Critical Value Backend
}
