% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvalue.R
\name{cvalue_sn}
\alias{cvalue_sn}
\title{Stage 1 Critical Value}
\usage{
cvalue_sn(x_data, alpha)
}
\arguments{
\item{x_data}{an n x k matrix of evaluated moment functions.}

\item{alpha}{the confidence level required.}
}
\value{
a float for the critical value.
}
\description{
Stage 1 Critical Value
}
