% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moment.R
\name{find_dist}
\alias{find_dist}
\title{Find Max Distance}
\usage{
find_dist(dist_data, j0_vector)
}
\arguments{
\item{dist_data}{an n x (J + 1) matrix of distances from the product
factories to the cities.}

\item{j0_vector}{a J0 x 2 matrix of ownership by the two firms.}
}
\value{
a n x 2 matrix of the maximum distance for each firm to each market.
}
\description{
Find Max Distance
}
