% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_bootstrap_indices}
\alias{get_bootstrap_indices}
\title{Get Bootstrap Indices}
\usage{
get_bootstrap_indices(
  num_rows,
  bootstrap_replications = NULL,
  rng_seed = NULL,
  bootstrap_indices = NULL
)
}
\arguments{
\item{num_rows}{Number of rows in the data.}

\item{bootstrap_replications}{the number of bootstrap replications.
Required if bootstrap_indices is not specified.}

\item{rng_seed}{the seed for replication purposes. If not specified, the
seed is not set.}

\item{bootstrap_indices}{an integer vector of indices to use for the
bootstrap. If this is specified, bootstrap_replications and rng_seed will
be ignored. If this is not specified, bootstrap_replications is required.}
}
\value{
an integer vector of indices to use for the bootstrap.
}
\description{
Get Bootstrap Indices
}
